refactor: 투표 결과 실시간 집계 방식을 Redis로 마이그레이션 (#131)
